done
revamped financial analyst, and admin
