--- src/app/analyst/create-forecast/page.css
@@ -1,51 +1,0 @@
-.grey-bg {
-    background-color: lightgrey !important;
-    border: 1px dotted !important;
-}
-
-.green-type {
-    background-color: lightgreen !important;
-    border: 1px solid black !important;
-    border-right: none !important;
-}
-
-.green-input {
-    background-color: lightgreen !important;
-    border: 1px solid black !important;
-    min-width: 60px;
-    border-left: none !important;
-}
-
-.bold {
-    font-weight: bolder;
-}
-
-.bottom-border {
-    border-bottom: 1px solid !important;
-}
-
-.underline {
-    text-decoration: underline 1px solid black;
-}
-
-.font {
-    font-size: 12px;
-}
-
-.title {
-    background-color: lightsteelblue !important;
-    font-size:smaller;
-}
-
-.button {
-    font-size: 10px !important;
-}
-
-.left {
-    text-align: left;
-}
-
-.right {
-    text-align: right;
-    font-style: italic;
-}
